remove lib.dms
